refactor(naming): rename CardTemplate to CardCore
